Setup DynamicLinks::ShortenedUrl model: - move DynamicLinks::ShortenedUrlV2 model code to DynamicLinks::ShortenedUrl model - remove DynamicLinks::ShortenedUrlV1 and DynamicLinks::ShortenedUrlV2 models - remove unused code on DynamicLinks module
